Backend/requirements-ubuntu.txt
# Ubuntu-compatible requirements file
# Based on successfully installed packages

# Core Django Framework
Django==3.2.25
djangorestframework==3.14.0
djangorestframework_simplejwt==4.4.0
django-cors-headers==3.10.1
django-ninja==0.19.1

# FastAPI and ASGI
fastapi==0.83.0
uvicorn==0.17.0
starlette==0.27.0

# Database
psycopg2-binary==2.9.7

# Task Queue and Scheduling  
celery==5.2.7
redis==4.6.0
APScheduler==3.9.1
kombu==5.2.4
billiard==3.6.4
vine==5.0.0

# Authentication and Security
PyJWT==2.8.0
python-jose==3.3.0
passlib==1.7.4
cryptography==41.0.7
bcrypt==4.0.1

# Computer Vision (Ubuntu compatible versions)
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
pillow==10.0.1
numpy==1.24.3
imageio==2.31.4

# HTTP and API clients
requests==2.31.0
urllib3==1.26.16
httplib2==0.22.0
h11==0.14.0

# Data Processing and Export
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0
tablib==3.5.0
et_xmlfile==1.1.0

# Admin Interface Enhancements
django-jazzmin==2.6.0
django-admin-honeypot==1.1.0
django-admin-logs==1.0.2
django-import-export==3.2.0
django-object-actions==4.2.0

# API Documentation
drf-yasg==1.21.7
uritemplate==4.1.1

# Google Services
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
oauthlib==3.2.2

# Communication
pyTelegramBotAPI==4.14.0

# Video Processing
ffmpeg-python==0.2.0
pytube==15.0.0

# Web Automation (Ubuntu compatible)
selenium==4.15.0
webdriver-manager==4.0.1

# Environment and Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Utilities and Dependencies
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
packaging==23.2
sqlparse==0.4.4

# Async and Networking
anyio==3.7.1
sniffio==1.3.0
websocket-client==1.6.4

# Text Processing and CLI
click==8.1.7
colorama==0.4.6
wcwidth==0.2.8

# Development and Compatibility
typing_extensions==4.8.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
defusedxml==0.7.1

# Pydantic and FastAPI dependencies
pydantic==2.4.2
annotated-types==0.6.0

# Additional utilities
python-multipart==0.0.6
diff-match-patch==20230430
